Using Map for Opts
Motivation: To pattern match the condition option match to more cleanly
ignore the to field

Quoting from Erlang OTP team => https://erlangforums.com/t/handling-options-in-erlang-otp-apis/1133
Briefly, the main motivation for avoiding lists unless necessary, is that given the option list [{use_foo,true}, {use_foo,false}] the value of the use_foo is not immediately clear. Existing OTP APIs that use options are not consistent in that regard. With maps, that ambiguity cannot arise.